refactor config stuff
