+++ multicamera_acquisition/config/__init__.py
