+++ multicamera_acquisition/config/default_azure_config.py
