+++ multicamera_acquisition/config/default_basler_config.py
