+++ multicamera_acquisition/config/default_display_config.py
