+++ multicamera_acquisition/config/default_ffmpeg_writer_config.py
