+++ multicamera_acquisition/config/default_nvc_writer_config.py
